pyproject.toml
[tool.poetry]
name = "playwright-pool"
version = "0.1.0"
description = ""
authors = ["tgscan.xyz@outlook.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-asyncio = "^0.23.6"
playwright = "^1.43.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
